表单模板创建

+++ forms-core/src/main/java/me/cxis/forms/model/FormQuestionVO.java
@@ -4,35 +4,35 @@ import java.io.Serializable;
 
 public class FormQuestionVO implements Serializable {
 
-    private int id;
+    private Long id;
 
-    private int formId;
+    private Long formId;
 
     private String title;
 
     private String description;
 
-    private int order;
+    private Integer order;
 
-    private int pageNumber;
+    private Integer pageNumber;
 
-    private int widgetType;
+    private Integer widgetType;
 
     private WidgetRuleVO widgetRule;
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public int getFormId() {
+    public Long getFormId() {
         return formId;
     }
 
-    public void setFormId(int formId) {
+    public void setFormId(Long formId) {
         this.formId = formId;
     }
 
@@ -52,27 +52,27 @@ public class FormQuestionVO implements Serializable {
         this.description = description;
     }
 
-    public int getOrder() {
+    public Integer getOrder() {
         return order;
     }
 
-    public void setOrder(int order) {
+    public void setOrder(Integer order) {
         this.order = order;
     }
 
-    public int getPageNumber() {
+    public Integer getPageNumber() {
         return pageNumber;
     }
 
-    public void setPageNumber(int pageNumber) {
+    public void setPageNumber(Integer pageNumber) {
         this.pageNumber = pageNumber;
     }
 
-    public int getWidgetType() {
+    public Integer getWidgetType() {
         return widgetType;
     }
 
-    public void setWidgetType(int widgetType) {
+    public void setWidgetType(Integer widgetType) {
         this.widgetType = widgetType;
     }
